requirements.txt
Flask==0.10.1
couchdb==1.0.1
python-swiftclient==3.0.0
python-keystoneclient